+++ client/src/App.js
@@ -1,13 +1,22 @@
+import './App.css'
+import Archeader from './components/Archeader'
+import Cabinet from './components/Cabinet'
+import Touchcontrols from './components/Touchcontrols';
 import Background from "./components/background";
-import {Router} from "react router";
+// import {Router} from "react router";
+
+
 function App() {
   return (
-    <Background>
-      <Router>
-        
-      </Router>
-    </Background>
-    
+  <div className="App">
+    {/* <Background> */}
+      {/* <Router>    */}
+      <Archeader/>
+       <Cabinet/>
+       {/* <Touchcontrols/> */}
+      {/* </Router> */}
+    {/* </Background> */}
+    </div>
   );
 }
