+++ client/src/components/Game.js
@@ -1,6 +1,6 @@
 import React from 'react'
 import Image from 'react-bootstrap/Image'
-import Pacman from './images/Pacresize.gif'
+import Pacman from '../images/Pacresize.gif'
 
 export default function Game() {
     return (
