+++ client/src/index.css
@@ -1,8 +1,84 @@
-* {
+/* * {
   margin: 0;
   padding: 0;
   background-image: url('./arcademachinebackground.jpeg');
   background-size: cover;
   height: 100%;
   width: 100%;
+} */
+
+body {
+  background-color:black;
+  margin: 0;
+  font-family: -apple-system, BlinkMacSystemFont, 'Segoe UI', 'Roboto', 'Oxygen',
+    'Ubuntu', 'Cantarell', 'Fira Sans', 'Droid Sans', 'Helvetica Neue',
+    sans-serif;
+  -webkit-font-smoothing: antialiased;
+  -moz-osx-font-smoothing: grayscale;
 }
+
+code {
+  font-family: source-code-pro, Menlo, Monaco, Consolas, 'Courier New',
+    monospace;
+}
+
+.container {
+  border:2px solid yellow;
+  
+
+}
+
+/* .snakegame{
+  border:1px solid yellow;
+} */
+
+
+iframe {
+  border-radius: 2%;
+  border: 0;
+  width: 73vw;
+  height: 33.2vw;
+  overflow: hidden;
+  margin-right: 1.5%;
+}
+
+
+.App {
+  text-align:center;
+}
+
+/* .cabinetbottom-fluid{
+  width:20%;
+  border:1px solid yellow;
+} */
+
+/* .arcadetop {
+ border:1px solid yellow;
+ margin:auto;
+ max-width:100%;
+}  */
+
+.arcadetop .img-fluid {
+  width:100vw;
+}
+
+.arcadebottom .img-fluid{
+  width:100vw;
+  
+}
+
+.gamevid .img-fluid {
+  width: 59vw;
+  position: absolute;
+  margin-left: -31%;
+  margin-top: 6%;
+}
+
+/* .arcadebottom {
+  border:1px solid yellow;
+  max-width:100%;
+  height:100vh;
+  
+} */
+
+
