--- src/App.css
@@ -1,164 +1,0 @@
-
-/* .Arcadetop{
-  border:1px solid yellow; 
-} */
-
-/* .App {
-  text-align: center; 
-}  */
-
-
-
-/* .thiscabinet-fluid {
-  max-width: 100%;  
-} */
-
-/* 
-@media (min-width:2048){
-  body {
-    background-size: 57%;
-    background-position: 399px -25%;
-    
-}
-.img-fluid {
-  max-width: 106%;
-  height: 10vw;
-  margin-left: -3%;
-}
-}
-
-
-
-
-
-
-@media (min-height:375px) and (max-height:376px){
-  .img-fluid {
-    max-width: 100%;
-    height: 14vw;
-    margin-left: 0%;
-    margin-top: -10%;
-}
-body {
-  background-size: 82%;
-  background-position: 65px -4%;
-}
-}
-
-
-
-
-
-@media (min-height:375px) and (max-height:374px){
-  .img-fluid {
-    max-width: 100%;
-    height: 13vw;
-    margin-left: 2%;
-    margin-top: -10%;
-}
-body {
-  background-size: 82%;
-  background-position: 65px -4%;
-}
-}
-
-
-
-
-@media (min-height:320px) and (max-height:375px){
-  .img-fluid {
-    max-width: 100%;
-    height: 13vw;
-    margin-left: 4%;
-    margin-top: -10%;
-}
-body {
-  background-size: 82%;
-  background-position: 65px -4%;
-}
-}
-
-
-
-
-@media (min-height:411px) and (max-height: 512px){
-  .img-fluid {
-      max-width: 100%;
-      height: 13vw;
-      margin-left: -1%;
-      margin-top: -10%;
-  }
-  body {
-    background-size: 82%;
-    background-position: 65px -4%;
-  }
-}
-  
-
-
-@media (min-height:340px) and (max-height:360px){
-  .img-fluid {
-    max-width: 100%;
-    height: 13vw;
-    margin-left: 1%;
-    margin-top: -10%;
-}
-body {
-  background-size: 82%;
-  background-position: 65px -4%;
-}
-}
-
-@media (min-width:2160px){
-  .img-fluid {
-    max-width: 100%;
-    height: 13vw;
-    margin-left: -1%;
-}
-body {
-  background-size: 80%;
-  background-position: 47% -24%;
-}
-}
-
-
-
-
-
-
-/* 
-.App-logo {
-  height: 40vmin;
-  pointer-events: none;
-}
-
-@media (prefers-reduced-motion: no-preference) {
-  .App-logo {
-    animation: App-logo-spin infinite 20s linear;
-  }
-}
-
-.App-header {
-  background-color: #282c34;
-  min-height: 100vh;
-  display: flex;
-  flex-direction: column;
-  align-items: center;
-  justify-content: center;
-  font-size: calc(10px + 2vmin);
-  color: white;
-  border:1px solid yellow;
-}
-
-.App-link {
-  color: #61dafb;
-}
-
-@keyframes App-logo-spin {
-  from {
-    transform: rotate(0deg);
-  }
-  to {
-    transform: rotate(360deg);
-  }
-} */
