--- src/App.js
@@ -1,21 +1,0 @@
-import './App.css'
-import Archeader from './Archeader'
-import Cabinet from './Cabinet'
-import Touchcontrols from './Touchcontrols';
-
-
-
-
-
-
-function App() {
-  return (
-    <div className="App">
-      <Archeader/>
-       <Cabinet/>
-       {/* <Touchcontrols/> */}
-  </div>
-  );
-}
-
-export default App;
