--- src/Snakegame.js
@@ -1,12 +1,0 @@
-import React from "react";
-
-
-const Snakegame = () => {
-  return (
-    <div className="snakegame">
-      <iframe scrolling = "no" title="games" src="http://vsnake.herokuapp.com"></iframe>
-    </div>
-  );
-}
-
-export default Snakegame;
