--- src/index.css
@@ -1,79 +1,0 @@
-body {
-  background-color:black;
-  margin: 0;
-  font-family: -apple-system, BlinkMacSystemFont, 'Segoe UI', 'Roboto', 'Oxygen',
-    'Ubuntu', 'Cantarell', 'Fira Sans', 'Droid Sans', 'Helvetica Neue',
-    sans-serif;
-  -webkit-font-smoothing: antialiased;
-  -moz-osx-font-smoothing: grayscale;
-}
-
-code {
-  font-family: source-code-pro, Menlo, Monaco, Consolas, 'Courier New',
-    monospace;
-}
-
-.container {
-  border:2px solid yellow;
-  
-
-}
-
-/* .snakegame{
-  border:1px solid yellow;
-} */
-
-::-webkit-scrollbar {
-  width: 0px;
-  height: 0px;
-}
-
-iframe {
-  border-radius: 2%;
-  border: 0;
-  width: 73vw;
-  height: 33.2vw;
-  overflow: hidden;
-  margin-right: 1.5%;
-}
-
-
-.App {
-  text-align:center;
-}
-
-/* .cabinetbottom-fluid{
-  width:20%;
-  border:1px solid yellow;
-} */
-
-/* .arcadetop {
- border:1px solid yellow;
- margin:auto;
- max-width:100%;
-}  */
-
-.arcadetop .img-fluid {
-  width:100vw;
-}
-
-.arcadebottom .img-fluid{
-  width:100vw;
-  
-}
-
-.gamevid .img-fluid {
-  width: 59vw;
-  position: absolute;
-  margin-left: -31%;
-  margin-top: 6%;
-}
-
-/* .arcadebottom {
-  border:1px solid yellow;
-  max-width:100%;
-  height:100vh;
-  
-} */
-
-
